+++ CHANGELOG.md
@@ -7,7 +7,18 @@ Nothing
 Nothing
 
 6.12.031: (Not released yet).
+[CHANGED]
+Now MultiBlock Heatexchanger has 3 testure to build.
+TungstenSteel/Tungsten/Tantalum Hafnium Carbide.
+They can output HU:16384/32768/131072.
 [ADDED]
-Multiblock Large Liuqid Burn Box.32768 HU/t split evenly over the 8 Heat Transmitter Outputs
+Multiblock Large Liquid Burn Box.Need liquid fuels to output HU.
+There 3 texture for build like heat exchanger,you can build them by TungstenSteel/Tungsten/Tantalum Hafnium Carbide.
+16384/32768/131072 HU/t split evenly over the 8 Heat Transmitter Outputs.
 
-*NOTE:For Original GT6 Verison changelogs.You can see in this https://gregtech.overminddl1.com/changelog 
+6.12.021:
+[CHANGED]Now Large Autoclave can accept energy.It can do 32 process at the same time too.
+[ADDED]
+Autoclave(LV-IV)That use energy to work. 
+
+*NOTE:For Original GT6 Version changelogs.You can see in this https://gregtech.overminddl1.com/changelog